14/n: Remove the use of groupId
GroupId was never used. FingerprintService#enroll sets groupId to
always be the same as userId. This change removes it from (almost)
everywhere.

We should update BiometricUtils and subclasses so that new HIDLs
do not take space for groupId. Since the framework persists the
info and we do not want to deal with database migration, let's
leave it there until the devices which shipped with groupId are
EOL.

Bug: 157790417

Test: Enroll fingerprints with a build before this change. Update
      to a build with this change - still able to authenticate, rename,
      remove previously enrolled fingerprints
Change-Id: I21391f5ffb71b27d7f9d4cfd4e6253b6001276dc