Add GET_INTENT_SENDER_INTENT permission to Shell
This is needed for a CTS test for picture-in-picture on TV.

As per go/privapp-permission:
1. This is the exact permission that guards specifically the
`PendingIntent::getIntent` method which is used in the test.

2. Two alternatives were considered:
a) Create another permisison and `@hide` it. However, this offers no
security benefit at the cost of maintenance overhead.
b) Do not create the test in the first place.

Bug: 153435517
Test: atest CtsSystemUiTestCases:PipNotificationTests
Change-Id: I04a5c82334a676351828fb82f2ac5886c888a0bc