Make the task affinity of a single-instance activity unique
The task affinity of a single-instance activity was the same as
other activities in the package by default. So, tasks that
have same task affinity would be removed from Recents when the
single-instance activity was started. Also, the trimmed activities
would be finished when return to home since 8ea5f08.

Bug: 166143360
Bug: 120237697
Test: atest RecentTasksTest
Change-Id: Ia903fa569239a55c0ea4587ac9e47547503a54e1