Keeps magnification button StartDelay for 3 secs even remove animations settings is on
When remove animations settings is on, all of the animation would be
out of function and so as the start delay of magnification button
animation. And this bug would make magnification mode unabled to be
changed through magnification button when the settings is on
because magnification button would be dismissed immediately when
starting to show the button.
To fix it, we use handler.postDelayed to handle the animation
start delay.

Bug: 167371292
Test: atest MagnificationModeSwitchTest
Change-Id: Ib4e3afc8fb4dab28283afb7826b4656e113f4723